+++ code/include/TransactionSuccessEvent.hpp
@@ -13,7 +13,11 @@ public:
 
   TransactionSuccessEvent();
 
-  TransactionSuccessEvent(ActorId ship_id, ActorId port_id, double ship_gold, double ship_rum, double port_rum);
+  TransactionSuccessEvent(ActorId ship_id,
+			  ActorId port_id,
+			  unsigned int ship_gold,
+			  unsigned int ship_rum,
+			  unsigned int port_rum);
 
   virtual ~TransactionSuccessEvent();
 
@@ -32,22 +36,22 @@ public:
   ActorId getPortId() const;
 
   /* Ship gold setter */
-  void setShipGold(double x);
+  void setShipGold(unsigned int x);
 
   /* Ship gold getter */
-  double getShipGold() const;
+  unsigned int getShipGold() const;
 
   /* Ship rum setter */
-  void setShipRum(double y);
+  void setShipRum(unsigned int y);
 
   /* Ship rum getter */
-  double getShipRum() const;
+  unsigned int getShipRum() const;
   
   /* Port rum setter */
-  void setPortRum(double y);
+  void setPortRum(unsigned int y);
 
   /* Port rum getter */
-  double getPortRum() const;
+  unsigned int getPortRum() const;
 
   /* TransactionSuccessEvent's event type */
   static const EventType event_type;
@@ -58,9 +62,9 @@ private:
   ActorId ship_id, port_id;
 
   /* Rum and gold of actors associated with this event */
-  double ship_gold;
-  double ship_rum;
-  double port_rum;
+  unsigned int ship_gold;
+  unsigned int ship_rum;
+  unsigned int port_rum;
 };
 
 inline EventType TransactionSuccessEvent::getEventType() const
@@ -88,32 +92,32 @@ inline ActorId TransactionSuccessEvent::getPortId() const
   return port_id;
 }
 
-inline void TransactionSuccessEvent::setShipGold(double x)
+inline void TransactionSuccessEvent::setShipGold(unsigned int x)
 {
   this->ship_gold = x;
 }
 
-inline double TransactionSuccessEvent::getShipGold() const
+inline unsigned int TransactionSuccessEvent::getShipGold() const
 {
   return ship_gold;
 }
 
-inline void TransactionSuccessEvent::setShipRum(double y)
+inline void TransactionSuccessEvent::setShipRum(unsigned int y)
 {
   this->ship_rum = y;
 }
 
-inline double TransactionSuccessEvent::getShipRum() const
+inline unsigned int TransactionSuccessEvent::getShipRum() const
 {
   return ship_rum;
 }
 
-inline void TransactionSuccessEvent::setPortRum(double y)
+inline void TransactionSuccessEvent::setPortRum(unsigned int y)
 {
   this->port_rum= y;
 }
 
-inline double TransactionSuccessEvent::getPortRum() const
+inline unsigned int TransactionSuccessEvent::getPortRum() const
 {
   return port_rum;
 }
